tests reorganized
