+++ src/impl_tuple.rs
@@ -230,121 +230,3 @@ impl_enumerable_for_tuple!(13, A, B, C, D, E, F, G, H, I, J, K, L, M);
 impl_enumerable_for_tuple!(14, A, B, C, D, E, F, G, H, I, J, K, L, M, N);
 impl_enumerable_for_tuple!(15, A, B, C, D, E, F, G, H, I, J, K, L, M, N, O);
 impl_enumerable_for_tuple!(16, A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P);
-
-#[cfg(test)]
-mod tests {
-    use crate::Enumerable;
-
-    fn collect_all<T: Enumerable>() -> Vec<T> {
-        T::enumerator().collect()
-    }
-
-    #[test]
-    fn test_tuple0() {
-        assert_eq!(vec![()], collect_all::<()>());
-    }
-
-    #[test]
-    fn test_tuple1() {
-        assert_eq!(vec![(false,), (true,)], collect_all::<(bool,)>());
-    }
-
-    #[test]
-    fn test_tuple2() {
-        // Illustrate the return order of the enumerator.
-        assert_eq!(
-            vec![(0, false), (0, true), (1, false), (1, true), (2, false)],
-            <(u8, bool)>::enumerator().take(5).collect::<Vec<_>>()
-        );
-
-        // Verify that the enumerator returns all possible values.
-        assert_eq!(
-            (0..=0xff)
-                .flat_map(|a| [false, true].into_iter().map(move |b| (a, b)))
-                .collect::<Vec<_>>(),
-            collect_all::<(u8, bool)>()
-        );
-    }
-
-    #[test]
-    fn test_tuple16() {
-        type Tuple16 = (
-            bool,
-            bool,
-            bool,
-            bool,
-            bool,
-            bool,
-            bool,
-            bool,
-            bool,
-            bool,
-            bool,
-            bool,
-            bool,
-            bool,
-            bool,
-            bool,
-        );
-        let tuple16_to_u16 = |t: Tuple16| {
-            let mut result = 0u16;
-            if t.0 {
-                result |= 1 << 15;
-            }
-            if t.1 {
-                result |= 1 << 14;
-            }
-            if t.2 {
-                result |= 1 << 13;
-            }
-            if t.3 {
-                result |= 1 << 12;
-            }
-            if t.4 {
-                result |= 1 << 11;
-            }
-            if t.5 {
-                result |= 1 << 10;
-            }
-            if t.6 {
-                result |= 1 << 9;
-            }
-            if t.7 {
-                result |= 1 << 8;
-            }
-            if t.8 {
-                result |= 1 << 7;
-            }
-            if t.9 {
-                result |= 1 << 6;
-            }
-            if t.10 {
-                result |= 1 << 5;
-            }
-            if t.11 {
-                result |= 1 << 4;
-            }
-            if t.12 {
-                result |= 1 << 3;
-            }
-            if t.13 {
-                result |= 1 << 2;
-            }
-            if t.14 {
-                result |= 1 << 1;
-            }
-            if t.15 {
-                result |= 1 << 0;
-            }
-            result
-        };
-
-        let collected_as_u16 = <Tuple16 as Enumerable>::enumerator()
-            .map(tuple16_to_u16)
-            .collect::<Vec<_>>();
-        assert_eq!(
-            collected_as_u16,
-            <u16 as Enumerable>::enumerator().collect::<Vec<_>>()
-        )
-    }
-}
